direction controller(advisor add hibernate exception)

+++ src/main/java/com/example/experienceexchange/service/interfaceService/IDirectionService.java
@@ -1,10 +1,8 @@
 package com.example.experienceexchange.service.interfaceService;
 
 import com.example.experienceexchange.dto.DirectionDto;
-import com.example.experienceexchange.model.Direction;
 
 import java.util.List;
-import java.util.Set;
 
 public interface IDirectionService {
     DirectionDto createDirection(DirectionDto directionDto);
@@ -15,5 +13,5 @@ public interface IDirectionService {
 
     List<DirectionDto> getAllDirections();
 
-    DirectionDto getDirection(Integer id);
+    DirectionDto getDirection(Long id);
 }
